double

+++ src/homography.py
@@ -170,7 +170,7 @@ def plane_grid(xbound, ybound, zs, yaws, rolls, pitchs):
     coords = rotation_matrix @ coords
     with open('master_coords_after.npy', 'wb') as f:
         np.save(f, coords.cpu().detach().numpy())
-    return coords.float()
+    return coords
 
 
 def ipm_from_parameters(image, xyz, K, RT, target_h, target_w, post_RT=None):
